Save work before manually deleting things. Up till now was claude assist
