+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/check/DestinationChecker.kt
@@ -5,29 +5,14 @@
 package io.airbyte.cdk.load.check
 
 import io.airbyte.cdk.load.command.Append
-import io.airbyte.cdk.load.command.DestinationCatalog
 import io.airbyte.cdk.load.command.DestinationConfiguration
 import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.command.NamespaceMapper
 import io.airbyte.cdk.load.data.ObjectTypeWithoutSchema
-import io.airbyte.cdk.load.message.DestinationRecordStreamComplete
-import io.airbyte.cdk.load.message.InputRecord
 import io.airbyte.cdk.load.schema.model.ColumnSchema
 import io.airbyte.cdk.load.schema.model.StreamTableSchema
 import io.airbyte.cdk.load.schema.model.TableName
 import io.airbyte.cdk.load.schema.model.TableNames
-import io.airbyte.cdk.load.util.serializeToString
-import io.airbyte.cdk.load.write.WriteOperation
-import io.github.oshai.kotlinlogging.KotlinLogging
-import java.io.InputStream
-import java.io.PipedInputStream
-import java.io.PipedOutputStream
-import java.io.PrintWriter
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-
-private val logger = KotlinLogging.logger {}
 
 /**
  * A check operation that is run before the destination is used.
@@ -66,50 +51,6 @@ interface DestinationChecker<C : DestinationConfiguration> {
 
     fun check(config: C)
     fun cleanup(config: C) {}
-}
-
-class DestinationCheckerSync<C : DestinationConfiguration>(
-    val catalog: DestinationCatalog,
-    stdinPipe: InputStream,
-    private val writeOperation: WriteOperation,
-    private val cleaner: CheckCleaner<C>,
-) : DestinationChecker<C> {
-    private val pipe: PrintWriter
-
-    init {
-        // See InputStreamProvider.make - in a CHECK operation, we swap the inputstream
-        // with a PipedInputStream.
-        val stdinPipeOutputStream = PipedOutputStream(stdinPipe as PipedInputStream)
-        pipe = PrintWriter(stdinPipeOutputStream)
-    }
-
-    override fun check(config: C) {
-        check(catalog.streams.size == 1) { "test catalog should have exactly 1 stream" }
-        val mockStream = catalog.streams.first()
-        runBlocking(Dispatchers.IO) {
-            launch { writeOperation.execute() }
-
-            pipe.println(
-                InputRecord(
-                        mockStream,
-                        """{"test": 42}""",
-                        System.currentTimeMillis(),
-                    )
-                    .asProtocolMessage()
-                    .serializeToString()
-            )
-            pipe.println(
-                DestinationRecordStreamComplete(mockStream, System.currentTimeMillis())
-                    .asProtocolMessage()
-                    .serializeToString()
-            )
-            pipe.close()
-        }
-    }
-
-    override fun cleanup(config: C) {
-        catalog.streams.forEach { stream -> cleaner.cleanup(config, stream) }
-    }
 }
 
 // TODO the cleaner maybe should also be looking for old test tables, a la DestinationCleaner??
