--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/pipeline/InputPartitioner.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.pipeline
-
-import io.airbyte.cdk.load.message.DestinationRecordRaw
-import io.micronaut.context.annotation.Secondary
-import jakarta.inject.Singleton
-
-/**
- * A dev interface for expressing how incoming data is partitioned. By default, data will be
- * partitioned by a hash of the stream name and namespace.
- */
-interface InputPartitioner {
-    fun getPartition(record: DestinationRecordRaw, numParts: Int): Int
-}
-
-/**
- * The default input partitioner, which partitions by the stream name. TODO: Should be round-robin?
- */
-@Singleton
-@Secondary
-class ByStreamInputPartitioner : InputPartitioner {
-    override fun getPartition(record: DestinationRecordRaw, numParts: Int): Int {
-        return Math.floorMod(record.stream.mappedDescriptor.hashCode(), numParts)
-    }
-}
