--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/pipeline/PipelineFlushStrategy.kt
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.pipeline
-
-interface PipelineFlushStrategy {
-    fun shouldFlush(
-        inputCount: Long,
-        dataAgeMs: Long,
-    ): Boolean
-}
