--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/pipeline/RandomInputPartitioner.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.pipeline
-
-import io.airbyte.cdk.load.message.DestinationRecordRaw
-import kotlin.random.Random
-
-class RandomInputPartitioner : InputPartitioner {
-    private val prng = Random(System.currentTimeMillis())
-
-    override fun getPartition(record: DestinationRecordRaw, numParts: Int): Int {
-        return prng.nextInt(numParts)
-    }
-}
