--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/task/Task.kt
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.task
-
-sealed interface TerminalCondition
-
-data object OnEndOfSync : TerminalCondition
-
-data object OnSyncFailureOnly : TerminalCondition
-
-data object SelfTerminating : TerminalCondition
-
-interface Task {
-    val terminalCondition: TerminalCondition
-
-    suspend fun execute()
-}
