+++ airbyte-integrations/connectors/destination-clickhouse/gradle.properties
@@ -1,2 +1,2 @@
-cdkVersion=0.1.105
+cdkVersion=local
 JunitMethodExecutionTimeout=10m
