+++ airbyte-integrations/connectors/destination-gcs-data-lake/gradle.properties
@@ -1,3 +1,3 @@
 testExecutionConcurrency=1
-cdkVersion=0.1.100
+cdkVersion=local
 JunitMethodExecutionTimeout=10m
