+++ airbyte-integrations/connectors/destination-postgres/build.gradle
@@ -7,7 +7,7 @@ plugins {
 
 airbyteBulkConnector {
     core = "load"
-    toolkits = ["load-object-storage"]
+    toolkits = ["load-csv"]
 }
 
 sourceSets {
